+++ src/main/java/com/thiagodev/springprojectbasic/Models/Cliente.java
@@ -21,7 +21,7 @@ public class Cliente implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String name;
 
     @Column(unique = true)
@@ -41,7 +41,7 @@ public class Cliente implements Serializable {
     @OneToMany(mappedBy="cliente",cascade = CascadeType.ALL)
     private List<Endereco> enderecos = new ArrayList<>();
 
-    public Cliente(Long id, String name, String email, String cpfOuCnpj, TipoCliente tipoCliente) {
+    public Cliente(Integer id, String name, String email, String cpfOuCnpj, TipoCliente tipoCliente) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -55,7 +55,7 @@ public class Cliente implements Serializable {
     public Cliente(Long id, String name, String email) {
     }
 
-    public Cliente(Object o, String name, String email, String cpfOuCnpj, Integer tipoCliente, String logradouro, String numero, String complemento, String bairro, String cep, String telefone1, String telefone2, String telefone3, Long cidadeId) {
+    public Cliente(Object o, String name, String email, String cpfOuCnpj, Integer tipoCliente, String logradouro, String numero, String complemento, String bairro, String cep, String telefone1, String telefone2, String telefone3, Integer cidadeId) {
     }
