+++ src/main/java/com/thiagodev/springprojectbasic/Models/Endereco.java
@@ -20,7 +20,7 @@ public class Endereco implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String logradouro;
     private String numero;
     private String complemento;
@@ -37,7 +37,7 @@ public class Endereco implements Serializable {
     private Cidade cidade;
 
     public Endereco(){}
-    public Endereco(Long id, String logradouro, String numero, String complemento, String bairro, String cep, Cliente cliente, Cidade cidade) {
+    public Endereco(Integer id, String logradouro, String numero, String complemento, String bairro, String cep, Cliente cliente, Cidade cidade) {
         this.id = id;
         this.logradouro = logradouro;
         this.numero = numero;
