+++ src/main/java/com/thiagodev/springprojectbasic/Models/Pagamento/PagamentoComBoleto.java
@@ -21,7 +21,7 @@ public class PagamentoComBoleto extends Pagamento {
     private Date dataPagamento;
 
 
-    public PagamentoComBoleto(Long id, EstadoPagamento estado, Pedido pedido, Date dataVencimento, Date dataPagamento) {
+    public PagamentoComBoleto(Integer id, EstadoPagamento estado, Pedido pedido, Date dataVencimento, Date dataPagamento) {
         super(id, estado, pedido);
         this.dataVencimento = dataVencimento;
         this.dataPagamento = dataPagamento;
