+++ src/main/java/com/thiagodev/springprojectbasic/Models/Pagamento/PagamentoComCartao.java
@@ -14,7 +14,7 @@ public class PagamentoComCartao extends Pagamento{
 
     private Integer numeroDeParcelas;  // se der erro é pq tem que criar o construtor através da superclasse
 
-    public PagamentoComCartao(Long id, EstadoPagamento estado, Pedido pedido, Integer numeroDeParcelas) {
+    public PagamentoComCartao(Integer id, EstadoPagamento estado, Pedido pedido, Integer numeroDeParcelas) {
         super(id, estado, pedido);
         this.numeroDeParcelas = numeroDeParcelas;
     }
