+++ src/main/java/com/thiagodev/springprojectbasic/Models/Produto.java
@@ -26,7 +26,7 @@ public class Produto implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String nome;
     private Double preco;
 
@@ -47,7 +47,7 @@ public class Produto implements Serializable {
         return lista;
     }
 
-    public Produto(Long id, String nome, Double preco) {
+    public Produto(Integer id, String nome, Double preco) {
         this.id = id;
         this.nome = nome;
         this.preco = preco;
