+++ src/main/java/com/thiagodev/springprojectbasic/repository/CidadeRepository.java
@@ -3,5 +3,5 @@ package com.thiagodev.springprojectbasic.repository;
 import com.thiagodev.springprojectbasic.Models.Cidade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CidadeRepository extends JpaRepository<Cidade, Long> {
+public interface CidadeRepository extends JpaRepository<Cidade, Integer> {
 }
