+++ src/main/java/com/thiagodev/springprojectbasic/repository/EnderecoRepository.java
@@ -3,5 +3,5 @@ package com.thiagodev.springprojectbasic.repository;
 import com.thiagodev.springprojectbasic.Models.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
 }
