+++ src/main/java/com/thiagodev/springprojectbasic/repository/EstadoRepository.java
@@ -3,5 +3,5 @@ package com.thiagodev.springprojectbasic.repository;
 import com.thiagodev.springprojectbasic.Models.Estado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoRepository extends JpaRepository<Estado, Long> {
+public interface EstadoRepository extends JpaRepository<Estado, Integer> {
 }
